refactor to toggleOrderPayed. Toggles payed state of order with customer in post-body
