+++ wwm/construction/meeting.go
@@ -10,7 +10,7 @@ import (
 	"github.com/WeisswurstSystems/WWM-BB/meeting/usecase/removeproduct"
 	"github.com/WeisswurstSystems/WWM-BB/meeting/usecase/setbuyer"
 	"github.com/WeisswurstSystems/WWM-BB/meeting/usecase/setplace"
-	"github.com/WeisswurstSystems/WWM-BB/meeting/usecase/payorder"
+	"github.com/WeisswurstSystems/WWM-BB/meeting/usecase/toggleorderpayed"
 )
 
 var MeetingStore = driver.NewMongoStore()
@@ -22,7 +22,7 @@ var MeetingUseCases = struct {
 	removeproduct.RemoveProductUseCase
 	setbuyer.SetBuyerUseCase
 	setplace.SetPlaceUseCase
-	payorder.PayOrderUsecase
+	toggleorderpayed.ToggleOrderPayedUseCase
 }{
 	createmeeting.Interactor{MeetingStore, UserUseCases.AuthenticateUseCase},
 	closemeeting.Interactor{MeetingStore, UserUseCases.AuthenticateUseCase},
@@ -30,7 +30,7 @@ var MeetingUseCases = struct {
 	removeproduct.Interactor{MeetingStore, UserUseCases.AuthenticateUseCase},
 	setbuyer.Interactor{MeetingStore, UserUseCases.AuthenticateUseCase},
 	setplace.Interactor{MeetingStore, UserUseCases.AuthenticateUseCase},
-	payorder.Interactor{MeetingStore, UserUseCases.AuthenticateUseCase},
+	toggleorderpayed.Interactor{MeetingStore, UserUseCases.AuthenticateUseCase},
 }
 
 var MeetingCommand = command.CommandHandler{
